% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{hrv_domain_vars}
\alias{hrv_domain_vars}
\title{HRV variable selector}
\format{
A named list containing 2 character vectors
\itemize{
\item \strong{time}: time domain variables
\item \strong{freq}: frequency domain variables
}
}
\usage{
hrv_domain_vars
}
\description{
list of Character to \code{dplyr::select()} HRV variables
}
\keyword{datasets}
