% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{HRV_report_vars}
\alias{HRV_report_vars}
\title{HRV report variables description}
\format{
A data.frame with the following columns:
\itemize{
\item \strong{variable}: variable name
\item \strong{description}: description of the variable
}
}
\usage{
HRV_report_vars
}
\description{
Description of variables for output of
}
\keyword{datasets}
