% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiolab-package.R
\docType{package}
\name{physiolab-package}
\alias{physiolab}
\alias{physiolab-package}
\title{physiolab: Functions for Laboratory Analysis}
\description{
Helper functions for wrangling laboratory data in physiology.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{kittipos.sir@mahidol.edu}

}
\keyword{internal}
