% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HRV_report.R
\name{read_HRV_report}
\alias{read_HRV_report}
\title{Read LabChart's HRV Report Text file(s)}
\usage{
read_HRV_report(
  file,
  ignore_missing_files = FALSE,
  text_field = NULL,
  docid_field = NULL,
  docvarsfrom = c("metadata", "filenames", "filepaths"),
  dvsep = "_",
  docvarnames = NULL,
  encoding = "UTF-16LE",
  source = NULL,
  cache = TRUE,
  format_cols = TRUE,
  ...
)
}
\arguments{
\item{file}{path to .txt files to be read. This is designed to automagically handle a number of common scenarios, so the value can be a "glob"-type wildcard value.}

\item{ignore_missing_files}{if \code{FALSE}, then if the file argument doesn't resolve to an existing file, then an error will be thrown. Note that this can happen in a number of ways, including passing a path to a file that does not exist, to an empty archive file, or to a glob pattern that matches no files.}

\item{text_field, docid_field}{a variable (column) name or column number indicating where to find the texts that form the documents for the corpus and their identifiers. This must be specified for file types .csv, .json, and .xls/.xlsx files. For XML files, an XPath expression can be specified.}

\item{docvarsfrom}{used to specify that docvars should be taken from the filenames, when the readtext inputs are filenames and the elements of the filenames are document variables, separated by a delimiter (\code{dvsep}).
This allows easy assignment of docvars from filenames such as 1789-Washington.txt, 1793-Washington, etc. by \code{dvsep} or from meta-data embedded in the text file header (headers). If docvarsfrom is set to "filepaths", consider the full path to the file, not just the filename.}

\item{dvsep}{separator (a regular expression character string) used in filenames to delimit docvar elements if \code{docvarsfrom="filenames"} or \code{docvarsfrom="filepaths"} is used}

\item{docvarnames}{character vector of variable names for \code{docvars}, if \code{docvarsfrom} is specified. If this argument is not used, default \code{docvar} names will be used (\code{docvar1}, \code{docvar2}, ...).}

\item{encoding}{vector: either the encoding of all files, or one encoding for each files}

\item{source}{used to specify specific formats of some input file types, such as JSON or HTML. Currently supported types are "twitter" for JSON and "nexis" for HTML}

\item{cache}{if TRUE, save remote file to a temporary folder. Only used when file is a URL.}

\item{format_cols}{If \code{TRUE} format appropriate output variables as factor, integer, or double. Compute \code{NN50_percent} (pNN50) from \code{NN50_count/Normals_count}}

\item{...}{passed to \code{readtext::readtext()}}
}
\value{
A data.frame
}
\description{
Parse LabChart's HRV Report .txt files in to a tidy tibble with one row per file.
Internally, this function calls \code{readtext::readtext()} and use regular expression to extract
information from each variables.
}
\examples{
# read_HRV_report("path/to/folder")
}
